+++ tests/testthat/test-ml.R
@@ -62,7 +62,7 @@ test_that("DL-MLP",{
     #skip_on_cran()
 
     model <- suppressWarnings(sits_train(samples_mt_ndvi, sits_deeplearning(
-        units = c(128,128), dropout_rates = c(0.5, 0.4), epochs = 50, verbose = 0)))
+        layers = c(128,128), dropout_rates = c(0.5, 0.4), epochs = 50, verbose = 0)))
 
     class.tb <- sits_classify(point_ndvi, model)
 
@@ -73,9 +73,8 @@ test_that("DL-MLP",{
 test_that("1D CNN model",{
     #skip_on_cran()
 
-    model <- suppressWarnings(sits_train(samples_mt_ndvi, sits_CNN(
-        units = c(32,32), kernels = c(9, 5),
-        dropout_rates = c(0.5, 0.4), epochs = 50, verbose = 0)))
+    model <- suppressWarnings(sits_train(samples_mt_ndvi, sits_FCN(layers = c(32,32), kernels = c(9, 5),
+                                                                   epochs = 50, verbose = 0)))
 
     class.tb <- sits_classify(point_ndvi, model)
 
@@ -86,10 +85,10 @@ test_that("1D CNN model",{
 test_that("tempCNN model",{
     #skip_on_cran()
 
-    model <- suppressWarnings(sits_train(samples_mt_ndvi, sits_tempCNN(
-        conv_units = c(32, 32), conv_kernels = c(7, 5),
-        conv_dropout_rates = c(0.5, 0.4),
-        node_units = c(128, 128), node_dropout_rates = c(0.5, 0.4),
+    model <- suppressWarnings(sits_train(samples_mt_ndvi, sits_TempCNN(
+        cnn_layers = c(32, 32), cnn_kernels = c(7, 5),
+        cnn_dropout_rates = c(0.5, 0.4),
+        mlp_layers = c(128), mlp_dropout_rates = c(0.5),
         epochs = 50, verbose = 0)))
 
     class.tb <- sits_classify(point_ndvi, model)
@@ -102,7 +101,7 @@ test_that("ResNet",{
     #skip_on_cran()
 
     model <- suppressWarnings(sits_train(samples_mt_ndvi, sits_ResNet(
-        units = 16, kernels = c(7,5,3), epochs = 50, verbose = 0)))
+        blocks = c(16, 16, 16), kernels = c(7,5,3), epochs = 50, verbose = 0)))
 
     class.tb <- sits_classify(point_ndvi, model)
 
@@ -111,3 +110,15 @@ test_that("ResNet",{
     expect_true(nrow(sits_show_prediction(class.tb)) == 16)
 })
 
+test_that("LSTM",{
+    #skip_on_cran()
+
+    model <- suppressWarnings(sits_train(samples_mt_ndvi, sits_LSTM_FCN(
+        cnn_layers = c(16, 16, 16), epochs = 50, verbose = 0)))
+
+    class.tb <- sits_classify(point_ndvi, model)
+
+    expect_true(all(class.tb$predicted[[1]]$class %in%
+                        sits_labels(samples_mt_ndvi)$label))
+    expect_true(nrow(sits_show_prediction(class.tb)) == 16)
+})