+++ tic.R
@@ -1,5 +1,0 @@
-do_package_checks()
-
-if (ci_on_travis()) {
-  do_pkgdown()
-}